requirements.txt
# Core data & analytics
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Market data
yfinance>=0.2.40

# LLM & agentic orchestration
langgraph>=0.0.40
openai>=1.30.0

# Streamlit (UI, added later but safe to include)
streamlit>=1.32.0

# PDF generation
fpdf>=1.7.2

# Utilities
python-dotenv>=1.0.1

# Testing
pytest>=8.0.0
